+++ src/FlowExplainer.Desktop/Diagnostics/LICS.cs
@@ -6,12 +6,12 @@ public class LICS : IGridDiagnostic
 {
 
     private IVectorField<Vec3, Vec2> VectorField;
-    public Func<Trajectory<Vec3>, float, float> Metric;
-    public float T = .2f;
-    public float Multi = 1f;
+    public Func<Trajectory<Vec3>, double, double> Metric;
+    public double T = .2f;
+    public double Multi = 1f;
     public IKernel Kernel = new ConstantKernel();
 
-    private static IVectorField<Vec2, float> NoiseField = new NoiseField();
+    private static IVectorField<Vec2, double> NoiseField = new NoiseField();
 
 
     public void UpdateGridData(GridVisualizer gridVisualizer)
@@ -27,8 +27,8 @@ public class LICS : IGridDiagnostic
             ref var atCoords = ref gridVisualizer.RegularGrid.AtCoords(new Vec2i(i, j));
             var pos = gridVisualizer.RegularGrid.ToWorldPos(new Vec2(i + .5f, j + .5f));
             var traj = flowop.Compute(t_start, t_end, pos, VectorField);
-            var totWeight = 0f;
-            var sum = 0f;
+            var totWeight = 0.0;
+            var sum = 0.0;
             foreach (var pair in traj.Enumerate())
             {
                 var weight = Kernel.GetWeight(pair.c);
